Make the cache directory user specific
This commit is kinda ugly because the cache directory is users
and ipso facto requires a system command to determine.

The directory is determined from the user environment ergo requires
shell expansion.

The shell expansion then creates an injection attack vector and
can not be combined with user inputs. Ipso facto, the full screenshot
path must be determined in two steps:
1) Get the base directory via a system command
2) Combine it with the id to generate the full path

To remove the ugliness, a `dump-screenshot` command should be
added to the `flight desktop` CLI