Setup the spec to use index_sessions and webport
There are two ports for each session: the vnc and websockify ports

The API needs to report the websockify port NOT the vnc port

However `flight desktop show` only reports the VNC port. Instead
`flight desktop list` reports both of them.

As this is a fairly major change, the spec has been updated in this
commit. Future commits will make the requisite changes in the code
base