Determine frequency of full refresh differently
Instead of dividing two integers and hoping for an integer result.
Let's multiply two integers ensuring we get an integer result.

The full refresh interval is no longer expressed as a number of seconds
which may or may not be honoured.  Instead it is defined as a multiple
of the short refresh interval, which will be honoured.